로그인_시에만_수정삭제_구현

+++ src/main/java/com/ndMo9/article/ArticleController.java
@@ -1,7 +1,11 @@
 package com.ndMo9.article;
 
+import com.ndMo9.user.SiteUser;
+import com.ndMo9.user.UserForm;
+import com.ndMo9.user.UserService;
 import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
+import org.springframework.http.HttpStatus;
 import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
@@ -10,7 +14,9 @@ import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.server.ResponseStatusException;
 
+import java.security.Principal;
 import java.util.List;
 
 @Controller
@@ -18,6 +24,7 @@ import java.util.List;
 @RequestMapping("/article")
 public class ArticleController {
     private final ArticleService articleService;
+    private final UserService userService;
 
     @GetMapping("/list")
     public String list(Model model) {
@@ -26,20 +33,21 @@ public class ArticleController {
         return "article_list";
     }
 
-    //  @PreAuthorize()
+    @PreAuthorize("isAuthenticated()")
     @GetMapping("/create")
     public String create(ArticleForm articleForm) {
 
         return "article_form";
     }
 
-    //  @PreAuthorize()
+    @PreAuthorize("isAuthenticated()")
     @PostMapping("/create")
-    public String create(@Valid ArticleForm articleForm, BindingResult bindingResult) {
+    public String create(@Valid ArticleForm articleForm, BindingResult bindingResult, Principal principal) {
         if (bindingResult.hasErrors()) {
             return "article_form";
         }
-        this.articleService.create(articleForm.getSubject(), articleForm.getContent());
+        SiteUser siteUser = this.userService.getUser(principal.getName());
+        this.articleService.create(articleForm.getSubject(), articleForm.getContent(), siteUser);
 
         return "redirect:/article/list";
     }
@@ -51,29 +59,38 @@ public class ArticleController {
         return "article_detail";
     }
 
-    //  @PreAuthorize()
+    @PreAuthorize("isAuthenticated()")
     @GetMapping("/modify/{id}")
-    public String modify(Model model, @PathVariable("id") Integer id) {
+    public String modify(Model model, @PathVariable("id") Integer id, UserForm userForm, Principal principal) {
         Article article = this.articleService.getArticle(id);
+        if(!article.getAuthor().getUsername().equals(principal.getName())) {
+            throw new ResponseStatusException(HttpStatus.BAD_REQUEST, "수정권한이 없습니다.");
+        }
         model.addAttribute("article", article);
         return "article_modify";
     }
 
-    //  @PreAuthorize()
+    @PreAuthorize("isAuthenticated()")
     @PostMapping("/modify/{id}")
-    public String modify(Model model, @PathVariable("id") Integer id, @Valid ArticleForm articleForm, BindingResult bindingResult) {
+    public String modify(Model model, @PathVariable("id") Integer id,Principal principal , @Valid ArticleForm articleForm, BindingResult bindingResult) {
         if (bindingResult.hasErrors()) {
             return "article_modify";
         }
         Article article = this.articleService.getArticle(id);
-
+        if (!article.getAuthor().getUsername().equals(principal.getName())) {
+            throw new ResponseStatusException(HttpStatus.BAD_REQUEST, "수정권한이 없습니다.");
+        }
         this.articleService.modify(article, articleForm.getSubject(), articleForm.getContent());
         return String.format("redirect:/article/detail/%s", id);
     }
 
-    //  @PreAuthorize()
+    @PreAuthorize("isAuthenticated()")
     @GetMapping("/delete/{id}")
-    public String delete(@PathVariable("id") Integer id) {
+    public String delete(@PathVariable("id") Integer id, Principal principal) {
+        Article article =  this.articleService.getArticle(id);
+        if (!article.getAuthor().getUsername().equals(principal.getName())) {
+            throw new ResponseStatusException(HttpStatus.BAD_REQUEST, "삭제권한이 없습니다.");
+        }
         this.articleService.delete(id);
         return "redirect:/article/list";
     }
